# synthesus_framework/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synthesus 2.0",
    "version": "2.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "post": {
        "summary": "Query",
        "operationId": "query_query_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/stream": {
      "get": {
        "summary": "Stream",
        "operationId": "stream_stream_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/characters": {
      "get": {
        "summary": "Characters",
        "operationId": "characters_characters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/process": {
      "post": {
        "summary": "Process",
        "operationId": "process_process_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/world_state": {
      "get": {
        "summary": "Get World State",
        "description": "Get current world state flags.",
        "operationId": "get_world_state_world_state_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Set World State",
        "description": "Set world state flags for all cognitive engines.",
        "operationId": "set_world_state_world_state_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Flags"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cognitive_stats": {
      "get": {
        "summary": "Cognitive Stats",
        "description": "Get statistics from all active cognitive engines.",
        "operationId": "cognitive_stats_cognitive_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "q": {
            "type": "string",
            "title": "Q",
            "default": ""
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          },
          "context": {
            "type": "string",
            "title": "Context",
            "default": ""
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "left"
          },
          "character": {
            "type": "string",
            "title": "Character",
            "default": ""
          },
          "player_id": {
            "type": "string",
            "title": "Player Id",
            "default": "default"
          }
        },
        "type": "object",
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "module": {
            "type": "string",
            "title": "Module"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "kernel"
          },
          "character": {
            "type": "string",
            "title": "Character",
            "default": ""
          },
          "pattern_id": {
            "type": "string",
            "title": "Pattern Id",
            "default": ""
          },
          "emotion": {
            "type": "string",
            "title": "Emotion",
            "default": ""
          },
          "relationship": {
            "additionalProperties": true,
            "type": "object",
            "title": "Relationship",
            "default": {}
          },
          "debug": {
            "additionalProperties": true,
            "type": "object",
            "title": "Debug",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "response",
          "confidence",
          "module"
        ],
        "title": "QueryResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}